[MIG] iot_input_data: Migration to 14.0
