+++ iot_input_data/models/__init__.py
@@ -1,1 +1,2 @@
 from . import iot_input_data
+from . import iot_device_input
